+++ package.json
@@ -14,8 +14,8 @@
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.22.2",
-		"@sveltejs/adapter-auto": "1.0.0-next.64",
-		"@sveltejs/kit": "1.0.0-next.405",
+		"@sveltejs/adapter-auto": "1.0.0-next.66",
+		"@sveltejs/kit": "1.0.0-next.431",
 		"@types/cookie": "^0.5.1",
 		"@typescript-eslint/eslint-plugin": "^5.27.0",
 		"@typescript-eslint/parser": "^5.27.0",
