+++ src/routes/auth/signin/+page.ts
@@ -1,0 +1,1 @@
+export const prerender = false
